Added more sprites and tried to fix an issue
So I added some dungeon tiles that i wanted to try and play around with to see if that looked any better, didn't really but I still have them in there. The other thing was that I've been getting this error that says "Null Reference Exception" when the player goes past the first level. I thought that this was because of the Player and GameManager being in the hierarchy rather than having the prefabs be variables of the loader but I guess I was wrong, need to look more into that because it seems like something he'd definitely point out when we present the final game